@@ -11,18 +11,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.1",
+    "@angular-devkit/build-angular": "^0.803.17",
+    "@angular/animations": "^8.2.13",
     "@angular/cdk": "^7.0.1",
-    "@angular/common": "^7.0.1",
-    "@angular/compiler": "^7.0.1",
-    "@angular/core": "^7.0.1",
-    "@angular/forms": "^7.0.1",
+    "@angular/common": "^8.2.13",
+    "@angular/compiler": "^8.2.13",
+    "@angular/core": "^8.2.13",
+    "@angular/forms": "^8.2.13",
     "@angular/http": "^7.0.1",
-    "@angular/platform-browser": "^7.0.1",
-    "@angular/platform-browser-dynamic": "^7.0.1",
+    "@angular/platform-browser": "^8.2.13",
+    "@angular/platform-browser-dynamic": "^8.2.13",
     "@angular/pwa": "^0.13.8",
-    "@angular/router": "^7.0.1",
-    "@angular/service-worker": "^7.0.1",
+    "@angular/router": "^8.2.13",
+    "@angular/service-worker": "^8.2.13",
     "@ng-bootstrap/ng-bootstrap": "^4.2.1",
     "angular2-cookie-law": "^6.0.4",
     "angular2-csv": "^0.2.9",
@@ -33,18 +34,18 @@
     "file-saver": "^2.0.2",
     "material-design-icons": "^3.0.1",
     "ng2-charts": "^2.2.2",
-    "ngx-bootstrap": "^3.3.0",
+    "ngx-bootstrap": "^5.2.0",
     "ngx-pagination": "^3.2.0",
     "ngx-smart-modal": "^7.1.0",
     "ngx-spinner": "^6.1.2",
-    "rxjs": "^6.3.3",
-    "zone.js": "~0.8.26"
+    "rxjs": "^6.5.3",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular/cli": "~6.1.4",
-    "@angular/compiler-cli": "^7.2.6",
-    "@angular/language-service": "^7.0.1",
+    "@angular/cli": "~8.3.17",
+    "@angular/compiler-cli": "^8.2.13",
+    "@angular/language-service": "^8.2.13",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -60,6 +61,6 @@
     "protractor": "^5.4.2",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~3.1.3"
+    "typescript": "~3.5.3"
   }
 }
